pyproject.toml
[tool.poetry]
name = "discord-py-template"
version = "0.0.0"
description = "A Discord.py bot template"
authors = ["Milo Weinberg <iapetus011@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
"discord.py" = "^2.0.0"
pydantic = "^1.9.2"
python-dotenv = "^0.20.0"
aiosqlite = "^0.17.0"
arrow = "^1.2.2"

[tool.poetry.dev-dependencies]
black = "^22.6.0"
isort = "^5.10.1"
flake8 = "^5.0.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
[tool.black]
line-length = 100
target-version = ['py310']
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 100
known_bot = ["bot"]
sections = ["FUTURE", "STDLIB", "THIRDPARTY", "BOT", "FIRSTPARTY", "LOCALFOLDER"]
